Restructure plugin architecture
