+++ src/admin/library/Free/Joomla/Table/Email.php
@@ -25,7 +25,7 @@ class Email extends BaseTable
     {
         parent::__construct('#__osdownloads_emails', 'id', $db);
 
-        MailingLists::loadObservers($this);
+        MailingLists\Manager::loadObservers($this);
     }
 
     public function store($updateNulls = false)
